test: move HTML minimizer tests into test/minify-option.test.js
Fold the 10 HTML-minimizer cases (and their snapshots) into the
existing `minify option` describe block in `test/minify-option.test.js`,
alongside the JS, JSON, and esbuild cases that already live there. The
standalone `test/htmlMinimizers.test.js` and its dedicated snapshot
file are removed.

The HTML fixtures (`fixtures/file.html`, `fixtures/fragment.html`,
`fixtures/html.js`, `fixtures/html-fragment.js`) stay where they are
and are now consumed from `minify-option.test.js`.
